feat:fix bug

+++ go_api_admin/app/march_voiced/models/bo/user.go
@@ -4,7 +4,7 @@ type SelectUserInfo struct {
 	Signature    string `json:"signature"`
 	FollowTotal  int64  `json:"followTotal"`
 	FansTotal    int64  `json:"fansTotal"`
-	ArticleTotal int    `json:"articleTotal"`
+	ArticleTotal int64  `json:"articleTotal"`
 	CollectTotal int64  `json:"collectTotal"`
 	UserIn
 }
